refactor(frontend): the workflow-reference bridge moves into @agenta/entity-ui
OSSdrillInUIProvider carried 588 lines of workflow-reference wiring that was not
OSS-specific — /m needs the same behaviour to render a config pane, and could not
reach it. It becomes useWorkflowReferenceBridge in the package, and the provider
keeps only the host's own composition.

The bridge imports nothing from @/oss, so it holds in a host that cannot see the
app layer.

Lane: this belongs with the entity-ui work, above pkg/session-surfaces — see
docs/design/sessions-ux-stack/restack-onto-112.md.
