+++ web/packages/agenta-entity-ui/src/DrillInView/index.ts
@@ -125,6 +125,9 @@ export type {
     WorkflowConfigPayload,
 } from "@agenta/ui/drill-in"
 
+// The workflow-as-tool reference bridge every host feeds into its own DrillInUIProvider.
+export {useWorkflowReferenceBridge} from "./bridges/useWorkflowReferenceBridge"
+
 // Core Types
 export type {
     // Data types
